+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Latest
 ------
+
+0.6.0 (2025-01-23)
+------------------
 * Added flag to APILoader so checklists can be updated even if the edited
   timestamp is unchanged.